Schema: enhance error messages
