+++ .changeset/ten-buckets-drive.md
@@ -1,0 +1,5 @@
+---
+"@effect/schema": patch
+---
+
+enhance error messages
