+++ packages/schema/test/Arbitrary/Arbitrary.test.ts
@@ -13,7 +13,7 @@ describe("Arbitrary > Arbitrary", () => {
   it("should throw on declarations without annotations", () => {
     const schema = S.declare(isUnknown)
     expect(() => Arbitrary.make(schema)).toThrow(
-      new Error("cannot build an Arbitrary for a declaration without annotations")
+      new Error("cannot build an Arbitrary for a declaration without annotations (<declaration schema>)")
     )
   })
